% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{as.data.frame,Counts-method}
\alias{as.data.frame,Counts-method}
\title{Converts Counts instance into data.frame}
\usage{
\S4method{as.data.frame}{Counts}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{Counts}{object to be converted}
}
\value{
data.frame with columns aa: Amino Acid, codon: Codon code, rscu: RSCU x(i, j)
for a given codon.
}
\description{
Converts Counts instance into data.frame
}
\examples{
Cub('A DNA', PA14CDS[[1]]) |>
  rscu() |>
  as.data.frame() |>
  head()
}
