% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Converts Counts instance into data.frame}
\usage{
as.data.frame(object, method, window)
}
\arguments{
\item{Counts}{object to be converted}
}
\value{
data.frame with columns aa: Amino Acid, codon: Codon code, rscu: Counts x(i, j)
for a given codon.
}
\description{
Converts Counts instance into data.frame
}
\examples{
dat <- rscu(PAO14CDS[[1]])
as.data.frame(dat)
}
