% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxmat.R
\name{auxmat}
\alias{auxmat}
\title{Auxiliary matrix}
\usage{
auxmat(m, vettcsi, vettphi, a, b, c, d, e)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{vettcsi}{Vector of the feeling parameters of the Beta-Binomial distribution, with length equal to the number of observations}

\item{vettphi}{Vector of the overdispersion parameters of the Beta-Binomial distribution, with length equal to the number of observations}

\item{a}{Real number}

\item{b}{Real number}

\item{c}{Real number}

\item{d}{Real number}

\item{e}{Real number}
}
\description{
Returns an auxiliary matrix needed for computing the variance-covariance matrix of a CUBE model with covariates.
}
\references{
Iannario, M. (2014). Modelling Uncertainty and Overdispersion in Ordinal Data,
  \emph{ Communications in Statistics- Theory and Methods}, \bold{43}, 771--786 \cr
 Piccolo, D. (2014). Inferential issues on CUBE models with covariates, 
 \emph{Communications in Statistics. Theory and Methods}, \bold{44}, DOI: 10.1080/03610926.2013.821487
}
\keyword{internal}
