% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC.GEM}
\alias{BIC.GEM}
\title{S3 BIC method for class "GEM"}
\usage{
\method{BIC}{GEM}(object, ...)
}
\arguments{
\item{object}{An object of class "GEM"}

\item{...}{Other arguments}
}
\value{
BIC index for the fitted model.
}
\description{
S3 BIC method for objects of class \code{\link{GEM}}.
}
\seealso{
\code{\link{logLik}}, \code{\link{GEM}}
}
\keyword{package}
