% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{cai,Cub-method}
\alias{cai,Cub-method}
\title{Calculates CAI value for a given gene}
\usage{
\S4method{cai}{Cub}(object, ref)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\value{
A numeric value for CAI
}
\description{
Calculates CAI value for a given gene
}
\examples{
dat <- Cub('A DNA', PA14CDS[[1]])
cai(dat)
}
