% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{cai}
\alias{cai}
\title{Calculates CAI value for a given gene}
\usage{
cai(object, ...)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\value{
RAC representing Relative Adaptiveness of a Codon for a given Gene / Genome
}
\description{
Calculates CAI value for a given gene
}
\examples{
dat <- Cub(PAO14CDS[[1]])
cai(dat)
}
