% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2cub.R
\name{chi2cub}
\alias{chi2cub}
\title{Pearson \eqn{X^2} statistic}
\usage{
chi2cub(m, ordinal, W, pai, gama)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{W}{Matrix of covariates for the feeling component}

\item{pai}{Uncertainty parameter}

\item{gama}{Vector of parameters for the feeling component, with length equal to NCOL(W)+1 
to account for an intercept term (first enry of gama)}
}
\value{
It returns the following results in a list: 
\item{df}{Number of degrees of freedom}
\item{chi2}{Value of the Pearson fitting measure}
\item{dev}{Deviance indicator}
}
\description{
Compute the \eqn{X^2} statistic of Pearson for CUB models with one or two discrete
 covariates for the feeling component.
}
\examples{
data(univer)
m<-7
ordinal<-univer[,8]
W<-univer[,4]
pai<-0.3
gama<-c(0.1,0.7)
pearson<-chi2cub(m, ordinal, W, pai, gama)
degfree<-pearson$df
statvalue<-pearson$chi2
deviance<-pearson$dev
}
\references{
Tutz, G. (2012). \emph{Regression for Categorical Data}, Cambridge University Press, Cambridge
}
\keyword{htest}

