% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.GEM}
\alias{coef.GEM}
\title{S3 Method: coef for class "GEM"}
\usage{
\method{coef}{GEM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{GEM}}}

\item{...}{Other arguments}
}
\value{
ML estimates of parameters of the fitted GEM model.
}
\description{
S3 method: coef for objects of class \code{\link{GEM}}.
}
\details{
Returns estimated values of coefficients of the fitted model
}
\seealso{
\code{\link{GEM}}, \code{\link{summary}}
}
\keyword{package}
