% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat}
\alias{cormat}
\title{Correlation matrix for estimated model}
\usage{
cormat(object,digits=options()$digits)
}
\arguments{
\item{object}{An object of class "GEM"}

\item{digits}{Number of significant digits to be printed. Default is \code{options()$digits}}
}
\value{
Parameters correlation matrix for fitted GEM models.
}
\description{
Compute parameter correlation matrix for estimated model as returned by an object
 of class "GEM".
}
\seealso{
\code{GEM}, \code{vcov}
}
\keyword{models}

