% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{count}
\alias{count}
\title{Counts codons' occurrence in a given Cub instance}
\usage{
count(object, ...)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\value{
A named integer vector representing count of a given ccodon in a DNAString
}
\description{
Counts codons' occurrence in a given Cub instance
}
\examples{
dat <- Cub('A DNA', PA14CDS[[1]])
count(dat)
}
