% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Counts-class}
\alias{Counts-class}
\title{An S4 class to represent a Relative Synonymous Codon Usage counts for the given
DNAString}
\description{
An S4 class to represent a Relative Synonymous Codon Usage counts for the given
DNAString
}
\section{Slots}{

\describe{
\item{\code{data}}{A named list of numerics, representing RSCU for a given codon}
}}

