% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Cub-class}
\alias{Cub-class}
\title{An S4 class to represent a Codon Usage Bias of a given DNA / Set of DNAs}
\description{
An S4 class to represent a Codon Usage Bias of a given DNA / Set of DNAs
}
\section{Slots}{

\describe{
\item{\code{id}}{A length-one character vector, giving the identifier to your Cub}

\item{\code{data}}{A character vector of codons on position 1, 4, 7, ... etc}
}}

\examples{
myCub <- Cub('A cub', PA14CDS[[1]])
myCub
}
