% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub-package.R
\docType{package}
\name{cub-package}
\alias{cub}
\alias{cub-package}
\title{cub: Collection of tools for Codon Usage Bias analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dmytro Strunin \email{dmytro.strunin@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
