% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub00.R
\name{cub00}
\alias{cub00}
\title{Main function for CUB models without covariates}
\usage{
cub00(m, ordinal, maxiter, toler)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{maxiter}{Maximum number of iterations allowed for running the optimization algorithm}

\item{toler}{Fixed error tolerance for final estimates}
}
\value{
An object of the class "CUB"
}
\description{
Function to estimate and validate a CUB model without covariates for given ordinal responses.
}
\seealso{
\code{\link{CUB}}, \code{\link{probbit}}, \code{\link{probcub00}}, \code{\link{loglikCUB}}
}
\keyword{internal}
