% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{CubSet-class}
\alias{CubSet-class}
\title{An S4 class to represent a Codon Usage Bias of a given Set of DNAs}
\description{
An S4 class to represent a Codon Usage Bias of a given Set of DNAs
}
\section{Slots}{

\describe{
\item{\code{data}}{A list of character vectors of codons on position 1, 4, 7, ... etc
for the given DNAStringSet}
}}

