% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubshe.R
\name{cubshe}
\alias{cubshe}
\title{Main function for CUB models with a shelter effect}
\usage{
cubshe(m, ordinal, shelter, maxiter, toler)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{shelter}{Category corresponding to the shelter choice}

\item{maxiter}{Maximum number of iterations allowed for running the optimization algorithm}

\item{toler}{Fixed error tolerance for final estimates}
}
\value{
An object of the class "CUB"
}
\description{
Estimate and validate a CUB model with a shelter effect.
}
\references{
Iannario M. (2012). Modelling \emph{shelter} choices in a class of mixture models for ordinal responses,
\emph{Statistical Methods and Applications}, \bold{21}, 1--22
}
\keyword{#models}
\keyword{internal}
