% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubvisual.R
\name{cubvisual}
\alias{cubvisual}
\title{Plot an estimated CUB model}
\usage{
cubvisual(ordinal,...)
}
\arguments{
\item{ordinal}{Vector of ordinal responses}

\item{...}{Additional arguments to be passed to plot() and text()}
}
\value{
A plot of the estimated parameter vector \eqn{(\pi, \xi)} as a point in the parameter space.
}
\description{
Plotting facility for the CUB estimation of ordinal responses.
}
\details{
It represents an estimated CUB model as a point
 in the parameter space with some useful options.
}
\examples{
data(univer)
attach(univer)
cubvisual(global,xlim=c(0,0.5),ylim=c(0.5,1),cex=0.8)
}
\keyword{device}

