% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cush00.R
\name{cush00}
\alias{cush00}
\title{CUSH model without covariates}
\usage{
cush00(m, ordinal, shelter, makeplot, summary)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{shelter}{Category corresponding to the shelter choice}

\item{makeplot}{Logical: if TRUE, the algorithm returns a graphical plot comparing fitted 
probabilities and observed relative frequencies}

\item{summary}{Logical: if TRUE, summary results of the fitting procedure are displayed on screen}
}
\value{
An object of the class "CUSH"
}
\description{
Estimate and validate a CUSH model for given ordinal responses, without covariates.
}
\keyword{internal}

