% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaprob.R
\name{deltaprob}
\alias{deltaprob}
\title{Mean difference of a discrete random variable}
\usage{
deltaprob(m, prob)
}
\arguments{
\item{m}{Number of categories}

\item{prob}{Probability distribution of the random variable}
}
\description{
Compute the Gini mean difference of a discrete random variable.
}
\examples{
m<-7
prob<-c(0.04,0.04,0.05,0.10,0.21,0.32,0.24)
deltaprob(m,prob)
}
\keyword{univar}

