% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effecubecsi.R
\name{effecubecsi}
\alias{effecubecsi}
\title{Auxiliary function for the log-likelihood estimation of CUBE models with covariates 
only for the feeling component}
\usage{
effecubecsi(param, ordinal, W, m)
}
\arguments{
\item{param}{Vector of length equal to NCOL(W) + 3 whose entries are the initial parameters estimates}

\item{ordinal}{Vector of ordinal responses}

\item{W}{Matrix of the selected covariates for explaining the feeling component}

\item{m}{Number of ordinal categories}
}
\description{
Compute the opposite of the scalar function that is maximized when running the 
E-M algorithm for CUBE models with covariates only for the feeling component.
}
\keyword{internal}

