% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effecush.R
\name{effecush}
\alias{effecush}
\title{Auxiliary function for the log-likelihood estimation of CUSH models with covariates}
\usage{
effecush(paravec, esternocush, shelter, m)
}
\arguments{
\item{paravec}{Vector of the initial parameters estimates}

\item{esternocush}{Matrix binding together the vector of ordinal data and the matrix XX of explanatory
variables whose first column is a column of ones needed to consider an intercept term}

\item{shelter}{Category corresponding to the shelter choice}

\item{m}{Number of ordinal categories}
}
\description{
Compute the opposite of the loglikelihood function for CUSH models
with covariates to explain the shelter effect.
}
\details{
It is called as an argument for "optim" within CUSH function (when no covariate is included)
 as the function to minimize.
}
\keyword{internal}
