% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effeihg.R
\name{effeihg}
\alias{effeihg}
\title{Auxiliary function for the log-likelihood estimation of IHG models without covariates}
\usage{
effeihg(theta, m, freq)
}
\arguments{
\item{theta}{Initial estimate for the parameter of the IHG distribution}

\item{m}{Number of ordinal categories}

\item{freq}{Vector of the absolute frequency distribution of the ordinal responses}
}
\description{
Compute the opposite of the log-likelihood function for an IHG model without covariates.
}
\details{
It is called as an argument for "optim" within IHG function (when no covariate is specified)
as the function to minimize.
}
\keyword{internal}
