% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/effeihgcov.R
\name{effeihgcov}
\alias{effeihgcov}
\title{Auxiliary function for the log-likelihood estimation of IHG models with covariates}
\usage{
effeihgcov(nu, ordinal, U, m)
}
\arguments{
\item{nu}{Vector of the starting values for the parameters to be estimated, with length equal to
NCOL(U)+1 to account for an intercept term (first entry of \eqn{nu})}

\item{ordinal}{Vector of ordinal responses}

\item{U}{Matrix of the explanatory variables for the preference parameter \eqn{\theta}}

\item{m}{Number of ordinal categories}
}
\description{
Compute the opposite of the log-likelihood function for an IHG model with covariates
for the preference parameter.
}
\details{
It is called as an argument for "optim" within IHG function (with covariates)
 as the function to minimize.
}
\keyword{internal}

