% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellegecub.R
\name{ellegecub}
\alias{ellegecub}
\title{Log-likelihood function for gecub distribution}
\usage{
ellegecub(ordinal,Y,W,X,bet,gama,omega,shelter)
}
\arguments{
\item{ordinal}{Vector of ordinal responses}

\item{Y}{Matrix of selected covariates for explaining the uncertainty component, not including intercept}

\item{W}{Matrix of selected covariates for explaining the feeling component, not including intercept}

\item{X}{Matrix of selected covariates for explaining the shelter effect, not including intercept}

\item{bet}{Matrix of selected covariates for explaining the uncertainty component, not including intercept}

\item{gama}{Matrix of selected covariates for explaining the feeling component, not including intercept}

\item{omega}{Matrix of selected covariates for explaining the shelter effect, not including intercept}

\item{shelter}{Category corresponding to the shelter choice}
}
\description{
Log-likelihood function for gecub distribution
}
\keyword{internal}

