% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Normalized Gini heterogeneity index}
\usage{
gini(prob)
}
\arguments{
\item{prob}{Vector of  probability distribution or relative frequencies}
}
\description{
Compute the normalized Gini heterogeneity
index for a given discrete probability distribution.
}
\examples{
prob<-c(0.04,0.04,0.05,0.10,0.21,0.32,0.24)
gini(prob)
}
\seealso{
\code{\link{laakso}}
}
\keyword{univar}

