% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihg00.R
\name{ihg00}
\alias{ihg00}
\title{Main function for IHG models without covariates}
\usage{
ihg00(m, ordinal, makeplot, summary)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{makeplot}{Logical: if TRUE, the function returns a graphical plot comparing fitted probabilities
and observed relative frequencies}

\item{summary}{Logical: if TRUE, summary results of the fitting procedure are displayed on screen}
}
\value{
An object of the class "IHG"

An object of the class "IHG"
}
\description{
Estimate and validate an IHG model without covariates for given ordinal responses.
}
\details{
The optimization procedure is run via "optim", option method="Brent" for constrained optimization 
(lower bound = 0, upper bound=1).
}
\keyword{internal}

