% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihgcov.R
\name{ihgcov}
\alias{ihgcov}
\title{Main function for IHG models with covariates}
\usage{
ihgcov(m, ordinal, U, makeplot, summary)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{U}{Matrix of selected covariates for the preference parameter}

\item{makeplot}{Logical: if TRUE and if only one dichotomous covariate is included in the model, with levels (0,1), 
the function returns a graphical plot comparing the distributions of the responses conditioned to the value of the 
covariate}

\item{summary}{Logical: if TRUE, summary results of the fitting procedure are displayed on screen}
}
\value{
An object of the class "IHG"

An object of the class "IHG"
}
\description{
Estimate and validate an IHG model for given ordinal responses, with covariates to 
explain the preference parameter.
}
\details{
The optimization procedure is run via "optim", option method="Brent" for constrained optimization 
(lower bound = 0, upper bound=1).
}
\keyword{internal}

