% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laakso.R
\name{laakso}
\alias{laakso}
\title{Normalized Laakso and Taagepera heterogeneity index}
\usage{
laakso(m, prob)
}
\arguments{
\item{m}{Number of categories}

\item{prob}{Vector of length \eqn{m} of a probability or relative frequency distribution}
}
\description{
Compute the normalized Laakso & Taagepera heterogeneity index for a given 
discrete probability distribution.
}
\examples{
m<-7
prob<-c(0.04,0.04,0.05,0.10,0.21,0.32,0.24)
laakso(m,prob)
}
\seealso{
\code{\link{gini}}
}
\keyword{univar}

