% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.GEM}
\alias{logLik.GEM}
\title{logLik S3 Method for class "GEM"}
\usage{
\method{logLik}{GEM}(object, ...)
}
\arguments{
\item{object}{An object of class "GEM"}

\item{...}{Other arguments}
}
\value{
Log-likelihood at the final ML estimates for parameters of the fitted GEM model.
}
\description{
S3 method: logLik() for objects of class "GEM".
}
\seealso{
\code{\link{loglikCUB}}, \code{\link{loglikCUBE}}, \code{\link{GEM}},  \code{\link{loglikIHG}},
\code{\link{loglikCUSH}}, \code{\link{BIC}}
}
\keyword{package}
