% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikcubecsi.R
\name{loglikcubecsi}
\alias{loglikcubecsi}
\title{Log-likelihood function of CUBE model with covariates only for feeling}
\usage{
loglikcubecsi(m, ordinal, W, pai, gama, phi)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{W}{Matrix of covariates for explaining the feeling component}

\item{pai}{Uncertainty parameter}

\item{gama}{Vector of parameters for the feeling component, with length 
equal to NCOL(W) + 1 to account for an intercept term (first entry of gama)}

\item{phi}{Overdispersion parameter}
}
\description{
Compute the log-likelihood function of a CUBE model for ordinal data with subjects' 
covariates only for feeling.
}
\seealso{
internal
}
\keyword{internal}

