% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikcubshe.R
\name{loglikcubshe}
\alias{loglikcubshe}
\title{Log-likelihood of a CUB model with shelter effect}
\usage{
loglikcubshe(m, freq, pai1, pai2, csi, shelter)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{freq}{Vector of the absolute frequency distribution}

\item{pai1}{Mixing coefficient for the shifted Binomial component of the mixture distribution}

\item{pai2}{Mixing coefficient for the discrete Uniform component of the mixture distribution}

\item{csi}{Feeling parameter}

\item{shelter}{Category corresponding to the shelter choice}
}
\description{
Compute the log-likelihood of a CUB model with a shelter effect
for the given absolute frequency distribution.
}
\keyword{internal}
