% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikcush00.R
\name{loglikcush00}
\alias{loglikcush00}
\title{Log-likelihood function for a CUSH model without covariates}
\usage{
loglikcush00(m,ordinal,delta,shelter)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{delta}{Shelter parameter}

\item{shelter}{Category corresponding to the shelter choice}
}
\description{
Compute the log-likelihood function for a CUSH model 
without covariate for the given ordinal responses.
}
\seealso{
\code{\link{GEM}}
}
\keyword{internal}
