% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikihgcov.R
\name{loglikihgcov}
\alias{loglikihgcov}
\title{Log-likelihood function for the IHG model with covariates}
\usage{
loglikihgcov(m, ordinal, U, nu)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{U}{Matrix of selected covariates for explaining the preference parameter}

\item{nu}{Vector of coefficients for covariates, whose length equals NCOL(U)+1 to include
an intercept term in the model (first entry of nu)}
}
\description{
Compute the log-likelihood function for the IHG model
 with covariates to explain the preference parameter.
}
\seealso{
loglikIHG
}
\keyword{internal}
