% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PA14CDS}
\alias{PA14CDS}
\title{P. aeruginosa PA14 CDS data}
\format{
\subsection{'PAO1CDS}{

A DNAStringSet with 5572 dna sequences
\describe{
\item{names}{Gene ID}
\item{width}{Length of the sequence}
\item{seq}{DNA sequence of the gene}
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/assembly/GCF_000014625.1/}
}
\usage{
PA14CDS
}
\description{
CDS extracted from the full genome assembly of P. aeruginosa PA14, assembly
ASM1462v1 provided by NCBI. Release date 2013-10-06
}
\keyword{datasets}
