% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAO1CDS}
\alias{PAO1CDS}
\title{P. aeruginosa PAO1 CDS data}
\format{
\subsection{'PAO1CDS}{

A DNAStringSet with 5572 dna sequences
\describe{
\item{names}{Gene ID}
\item{width}{Length of the sequence}
\item{seq}{DNA sequence of the gene}
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/assembly/GCF_000006765.1/}
}
\usage{
PAO1CDS
}
\description{
CDS extracted from the full genome assembly of P. aeruginosa PAO1, assembly
ASM676v1 provided by NCBI. Release date 2006-07-07
}
\keyword{datasets}
