% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parnames.R
\name{parnames}
\alias{parnames}
\alias{parnames.GEM}
\title{Generic function for coefficient names}
\usage{
parnames(object)
}
\arguments{
\item{object}{An object of class "GEM"}
}
\description{
Generic function for names of parameter estimates of object of class "GEM".
}
\examples{
data(univer);attach(univer)
model<-GEM(Formula(officeho~0|0|0),family="cub",shelter=7)
model 
}
\seealso{
\code{\link{summary}}
}
\keyword{internal}
