% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{plot,CubSet-method}
\alias{plot,CubSet-method}
\title{Plots RSCU as a stucked bar chart, represented RSCU(i, j) for each codon, grouped by
amino acid.}
\usage{
\S4method{plot}{CubSet}(object, value = "count", scale = FALSE, ...)
}
\arguments{
\item{x}{A RSCU object to be plotted}
}
\value{
A ggplot2::ggplot object
}
\description{
Plots RSCU as a stucked bar chart, represented RSCU(i, j) for each codon, grouped by
amino acid.
}
\examples{
Cub('A DNA', PA14CDS[[1]]) |>
  rscu() |>
  plot()
}
