% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{plot}
\alias{plot}
\title{Plots Counts as a stucked bar chart, represented RSCU(i, j) for each codon, grouped by
amino acid.}
\usage{
plot(object, method, window)
}
\arguments{
\item{x}{A Counts object to be plotted}
}
\value{
A ggplot2::ggplot object
}
\description{
Plots Counts as a stucked bar chart, represented RSCU(i, j) for each codon, grouped by
amino acid.
}
\examples{
dat <- rscu(PAO14)
plot(dat)
}
