% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotloglikihg.R
\name{plotloglikihg}
\alias{plotloglikihg}
\title{Plot of the log-likelihood function of the IHG distribution}
\usage{
plotloglikihg(m,freq)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{freq}{Vector of the absolute frequency distribution}
}
\description{
Plot the log-likelihood function of an IHG model fitted to a given absolute frequency distribution,
 over the whole support of the preference parameter. It returns also the ML estimate.
}
\examples{
m<-7
freq<-c(828,275,202,178,143,110,101)
max<-plotloglikihg(m,freq)
}
\seealso{
\code{\link{loglikIHG}}
}
