% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print,Cub-method}
\alias{print,Cub-method}
\title{Prints of a Cub object.}
\usage{
\S4method{print}{Cub}(object, ...)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\description{
Prints of a Cub object.
}
\examples{
dat <- Cub('A DNA', PAO1CDS[[1]])
summary(dat)
}
