% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.GEM}
\alias{print.GEM}
\title{S3 method: print for class "GEM"}
\usage{
\method{print}{GEM}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{GEM}}}

\item{...}{Other arguments}
}
\value{
Brief summary results of the fitting procedure, including parameter estimates, their standard errors and 
the executed call.
}
\description{
S3 method print for objects of class \code{\link{GEM}}.
}
\keyword{package}
