% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{print}
\alias{print}
\title{Prints a Cub object.}
\usage{
print(object, ...)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\description{
Prints a Cub object.
}
\examples{
dat <- Cub(PAO14[[1]])
print(dat)
}
