% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2gecub.R
\name{Q2gecub}
\alias{Q2gecub}
\title{Auxiliary function for the log-likelihood estimation of GeCUB models.}
\usage{
Q2gecub(param,datidue)
}
\arguments{
\item{param}{Vector of initial estimates of parameters for the feeling component}

\item{datidue}{Auxiliary matrix}
}
\description{
Define the opposite one of the two scalar functions that are maximized when running the E-M algorithm
for GeCUB models with covariates for feeling, uncertainty and overdispersion.
}
\keyword{internal}
