% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Qdue.R
\name{Qdue}
\alias{Qdue}
\title{Auxiliary function for the log-likelihood estimation of CUBE models with covariates}
\usage{
Qdue(param, esterno2, q, m)
}
\arguments{
\item{param}{Vector of initial estimates of parameters for the feeling component and the overdispersion effect}

\item{esterno2}{Matrix binding together the column vector of the posterior probabilities that each observed rating
 has been generated by the distribution of the first component of the mixture, the column vector of ordinal responses,
and the matrices W and Z of the selected covariates for feeling and overdispersion, respectively}

\item{q}{Number of selected covariates for explaining the feeling component}

\item{m}{Number of ordinal categories}
}
\description{
Define the opposite of one of the two scalar functions that are maximized when running the E-M
algorithm for CUBE models with covariates for feeling, uncertainty and overdispersion.
}
\details{
It is iteratively called as an argument of "optim" within CUBE function (with covariates) as  the function to minimize
to compute the maximum likelihood estimates for the feeling and the overdispersion components.
}
\keyword{internal}

