% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quno.R
\name{Quno}
\alias{Quno}
\title{Auxiliary function for the log-likelihood estimation of CUBE models with covariates}
\usage{
Quno(bet, esterno1)
}
\arguments{
\item{bet}{Vector of initial estimates of parameters for the uncertainty component}

\item{esterno1}{Matrix binding together the column vector of the posterior probabilities
that each observed rating has been generated by the first component distribution of the mixture, with the matrix
YY of explicative  variables for the uncertainty component, expanded with a unitary vector in the first column to
consider also an intercept term}
}
\description{
Define the opposite one of the two scalar functions that are maximized when running the E-M algorithm
for CUBE models with covariates for feeling, uncertainty and overdispersion.
}
\details{
It is iteratively called as an argument of "optim" within CUBE function (with covariates) as  the function
to minimize  to compute the maximum likelihood estimates for the feeling and the overdispersion components.
}
\keyword{internal}
