% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qunogecub.R
\name{Qunogecub}
\alias{Qunogecub}
\title{Auxiliary function for the log-likelihood estimation of GeCUB models.}
\usage{
Qunogecub(param,datiuno,s)
}
\arguments{
\item{param}{Vector of initial estimates of parameters for the uncertainty component}

\item{datiuno}{Auxiliary matrix}

\item{s}{Number of covariates to explain the shelter effect}
}
\description{
Define the opposite one of the two scalar functions that are maximized when running the E-M algorithm
for GeCUB models with covariates for feeling, uncertainty and overdispersion.
}
\keyword{internal}

