% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{rac,Cub-method}
\alias{rac,Cub-method}
\title{Creates a Count instance from a given Cub object}
\usage{
\S4method{rac}{Cub}(object, ref)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\value{
RSCU representing Relative Synonymous Codon Usage for a given Gene / Genome
}
\description{
Creates a Count instance from a given Cub object
}
\examples{
dat <- Cub('A DNA', PA14CDS[[1]])
rscu(dat)
}
