% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{rac}
\alias{rac}
\title{Creates a RAC instance from a given Cub object}
\usage{
rac(object, ...)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\value{
RAC representing Relative Adaptiveness of a Codon for a given Gene / Genome
}
\description{
Creates a RAC instance from a given Cub object
}
\examples{
dat <- Cub(PAO14[[1]])
rac(dat)
}
