% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{rscu,Cub-method}
\alias{rscu,Cub-method}
\title{Creates a RSCU instance from a given Cub object}
\usage{
\S4method{rscu}{Cub}(object)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\value{
RSCU representing Relative Synonymous Codon Usage for a given Gene / Genome
}
\description{
Creates a RSCU instance from a given Cub object
}
\examples{
dat <- Cub('A DNA', PA14CDS[[1]])
rscu(dat)
}
