% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{rscu}
\alias{rscu}
\title{Creates a Counts instance from a given Cub object}
\usage{
rscu(object, ...)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\value{
Counts representing Relative Synonymous Codon Usage for a given Gene / Genome
}
\description{
Creates a Counts instance from a given Cub object
}
\examples{
dat <- Cub(PAO14[[1]])
rscu(dat)
}
