% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{show,Cub-method}
\alias{show,Cub-method}
\title{Shows a Cub object.}
\usage{
\S4method{show}{Cub}(object)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\description{
Shows a Cub object.
}
\examples{
dat <- Cub('A DNA', PAO1CDS[[1]])
dat
}
