% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simcube.R
\name{simcube}
\alias{simcube}
\title{Simulation routine for CUBE models}
\usage{
simcube(n,m,pai,csi,phi)
}
\arguments{
\item{n}{Number of simulated observations}

\item{m}{Number of ordinal categories}

\item{pai}{Uncertainty parameter}

\item{csi}{Feeling parameter}

\item{phi}{Overdispersion parameter}
}
\description{
Generate \eqn{n} pseudo-random observations following the given CUBE
distribution.
}
\examples{
n<-300
m<-9
pai<-0.7
csi<-0.4
phi<-0.1
simulation<-simcube(n,m,pai,csi,phi)
plot(table(simulation),xlab="Ordinal categories",ylab="Frequencies")
}
\seealso{
\code{\link{probcube}}
}
\keyword{distribution}

