% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simihg.R
\name{simihg}
\alias{simihg}
\title{Simulation routine for IHG models}
\usage{
simihg(n,m,theta)
}
\arguments{
\item{n}{Number of simulated observations}

\item{m}{Number of ordinal categories}

\item{theta}{Preference parameter}
}
\description{
Generate \eqn{n} pseudo-random observations following the given IHG distribution.
}
\examples{
n<-300
m<-9
theta<-0.4
simulation<-simihg(n,m,theta)
plot(table(simulation),xlab="Number of categories",ylab="Frequencies")
}
\seealso{
\code{\link{probihg}}
}
\keyword{distribution}
