% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary,Cub-method}
\alias{summary,Cub-method}
\title{Shows summary of a Cub object.}
\usage{
\S4method{summary}{Cub}(object, ...)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\description{
Shows summary of a Cub object.
}
\examples{
dat <- Cub('A DNA', PAO1CDS[[1]])
summary(dat)
}
