% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary,CubSet-method}
\alias{summary,CubSet-method}
\title{Shows summary of a Cub object.}
\usage{
\S4method{summary}{CubSet}(object, ...)
}
\arguments{
\item{object}{A CubSet instance to summarize}
}
\description{
Shows summary of a Cub object.
}
\examples{
dat <- Cub('A DNA', PA14CDS[[1]])
summary(dat)
}
