% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{summary}
\alias{summary}
\title{Shows summary of a Cub object.}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{A Cub instance to summarize}
}
\description{
Shows summary of a Cub object.
}
\examples{
dat <- Cub(PAO14CDS[[1]])
summary(dat)
}
