% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcub00.R
\name{varcub00}
\alias{varcub00}
\title{Variance of CUB models without covariates}
\usage{
varcub00(m,pai,csi)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{pai}{Uncertainty parameter}

\item{csi}{Feeling parameter}
}
\description{
Compute the variance of a CUB model without covariates.
}
\examples{
m<-9
pai<-0.6
csi<-0.5
varcub<-varcub00(m,pai,csi)
}
\references{
Piccolo D. (2003). On the moments of a mixture of uniform and shifted binomial random variables. 
\emph{Quaderni di Statistica}, \bold{5}, 85--104
}
\seealso{
\code{\link{expcub00}}, \code{\link{probcub00}}
}
\keyword{distribution}
