% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.GEM}
\alias{vcov.GEM}
\title{S3 method  vcov() for class "GEM"}
\usage{
\method{vcov}{GEM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{GEM}}}

\item{...}{Other arguments}
}
\value{
Variance-covariance matrix of the final ML estimates for parameters of the fitted GEM model.
It returns the square of the estimated standard error for CUSH and IHG models with no covariates.
}
\description{
S3 method: vcov for objects of class \code{\link{GEM}}.
}
\seealso{
\code{\link{varmatCUB}}, \code{\link{varmatCUBE}}, \code{\link{GEM}}
}
\keyword{package}
