% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cushcov.R
\name{cushcov}
\alias{cushcov}
\title{CUSH model with covariates}
\usage{
cushcov(m, ordinal, X, shelter)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{X}{Matrix of selected covariates for explaining the shelter effect}

\item{shelter}{Category corresponding to the shelter choice}
}
\value{
An object of the class "GEM", "CUSH"
}
\description{
Estimate and validate a CUSH model for ordinal responses, with covariates
 to explain the shelter effect.
}
\keyword{internal}

