% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deltaprob.R
\name{deltaprob}
\alias{deltaprob}
\title{Mean difference of a discrete random variable}
\usage{
deltaprob(prob)
}
\arguments{
\item{prob}{Vector of the probability distribution}
}
\value{
Numeric value of the Gini mean difference of the input probability distribution,
 computed according to the de Finetti-Paciello formulation.
}
\description{
Compute the Gini mean difference of a discrete distribution
}
\examples{
prob<-c(0.04,0.04,0.05,0.10,0.21,0.32,0.24)
deltaprob(prob)
}
\keyword{univar}

