% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Normalized Gini heterogeneity index}
\usage{
gini(m, prob)
}
\arguments{
\item{m}{Number of categories}

\item{prob}{Vector of the probability distribution or relative frequencies}
}
\description{
Compute the normalized Gini heterogeneity 
index for a given discrete probability distribution.
}
\examples{
m<-7
prob<-c(0.04,0.04,0.05,0.10,0.21,0.32,0.24)
gini(m, prob)
}
\seealso{
\code{\link{laakso}}
}
\keyword{univar}

