% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ihg00.R
\name{ihg00}
\alias{ihg00}
\title{Main function for IHG models without covariates}
\usage{
ihg00(m, ordinal)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}
}
\value{
An object of the class "IHG"

An object of the class "IHG"
}
\description{
Estimate and validate an IHG model without covariates for given ordinal responses.
}
\details{
The optimization procedure is run via "optim", option method="Brent" for constrained optimization
(lower bound = 0, upper bound=1).
}
\keyword{internal}

