% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loglikCUB.R
\name{loglikCUB}
\alias{loglikCUB}
\title{Log-likelihood function for CUB models}
\usage{
loglikCUB(ordinal,m,param,Y=0,W=0,X=0,shelter=0)
}
\arguments{
\item{ordinal}{Vector of ordinal responses}

\item{m}{Number of ordinal categories}

\item{param}{Vector of parameters for the specified CUB model}

\item{Y}{Matrix of selected covariates to explain the uncertainty component (default: no covariate is included
in the model)}

\item{W}{Matrix of selected covariates to explain the feeling component (default: no covariate is included
in the model)}

\item{X}{Matrix of selected covariates to explain the shelter effect (default: no covariate is included
in the model)}

\item{shelter}{Category corresponding to the shelter choice (default: no shelter effect is included in the
model)}
}
\description{
Compute the log-likelihood value of a CUB model fitting given data, with or without covariates to
explain the feeling and uncertainty components, or for extended CUB models with shelter effect.
}
\details{
If no covariate is included in the model, then \code{param} should be given in the form \eqn{(\pi,\xi)}.
More generally, it should have the form \eqn{(\bold{\beta,\gamma)}} where,
respectively, \eqn{\bold{\beta}} and \eqn{\bold{\gamma}} are the vectors of
coefficients explaining the uncertainty and the feeling components, with length NCOL(Y)+1 and
 NCOL(W)+1 to account for an intercept term in the first entry. When shelter effect is considered, \code{param} corresponds
 to the first possibile parameterization and hence should be given as \code{(pai1,pai2,csi)}.
 No missing value should be present neither
  for \code{ordinal} nor for covariate matrices: thus, deletion or imputation procedures should be preliminarily run.
}
\examples{
## Log-likelihood of a CUB model with no covariate
m<-9; n<-300
pai<-0.6; csi<-0.4
ordinal<-simcub(n,m,pai,csi)
param<-c(pai,csi)
loglikcub<-loglikCUB(ordinal,m,param)
##################################
## Log-likelihood of a CUB model with covariate for uncertainty
\donttest{
data(relgoods)
m<-10
naord<-which(is.na(relgoods$Physician))
nacov<-which(is.na(relgoods$Gender))
na<-union(naord,nacov)
ordinal<-relgoods$Physician[-na]; Y<-relgoods$Gender[-na]
bbet<-c(-0.81,0.93); ccsi<-0.2
param<-c(bbet,ccsi)
loglikcubp0<-loglikCUB(ordinal,m,param,Y=Y)
#######################
## Log-likelihood of a CUB model with covariate for feeling
data(relgoods)
m<-10
naord<-which(is.na(relgoods$Physician))
nacov<-which(is.na(relgoods$Gender))
na<-union(naord,nacov)
ordinal<-relgoods$Physician[-na]; W<-relgoods$Gender[-na]
pai<-0.44; gama<-c(-0.91,-0.7)
param<-c(pai,gama)
loglikcub0q<-loglikCUB(ordinal,m,param,W=W)
#######################
## Log-likelihood of a CUB model with covariates for both parameters
data(relgoods)
m<-10
naord<-which(is.na(relgoods$Walking))
nacovpai<-which(is.na(relgoods$Gender))
nacovcsi<-which(is.na(relgoods$Smoking))
na<-union(naord,union(nacovpai,nacovcsi))
ordinal<-relgoods$Walking[-na]
Y<-relgoods$Gender[-na]; W<-relgoods$Smoking[-na]
bet<-c(-0.45,-0.48); gama<-c(-0.55,-0.43)
param<-c(bet,gama)
loglikcubpq<-loglikCUB(ordinal,m,param,Y=Y,W=W)
#######################
### Log-likelihood of a CUB model with shelter effect
m<-7; n<-400
pai<-0.7; csi<-0.16; delta<-0.15
shelter<-5
ordinal<-simcubshe(n,m,pai,csi,delta,shelter)
pai1<- pai*(1-delta); pai2<-1-pai1-delta
param<-c(pai1,pai2,csi)
loglik<-loglikCUB(ordinal,m,param,shelter=shelter)
##############
### Log-likelihood of a GeCUB
data(univer)
ordinal<-univer$officeho; Y<-W<-X<-univer$gender;
modelgecub<-GEM(Formula(ordinal~Y|W|X),family="cub",shelter=7,maxiter=100)
logLik(modelgecub)
param<-rep(0.1,6)
loglik<-loglikCUB(ordinal,m=7,param=param,shelter=7,Y=Y,W=W,X=X)
}
}
\seealso{
\code{\link{logLik}}
}
\keyword{htest}

