% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikihg_1.R
\name{loglikIHG}
\alias{loglikIHG}
\title{Log-likelihood function for IHG models}
\usage{
loglikIHG(ordinal,m,param,U=0)
}
\arguments{
\item{ordinal}{Vector of ordinal responses (factor type)}

\item{m}{Number of ordinal categories}

\item{param}{Vector of parameters for the specified IHG model}

\item{U}{Matrix of selected covariates to explain the preference parameter (default: no covariate is included 
in the model)}
}
\description{
Compute the log-likelihood function for IHG models with or without covariates 
to explain the preference parameter.
}
\details{
If no covariate is included in the model, then \code{param} is the estimate of the preference
parameter (theta), otherwise \code{param} has length equal to NCOL(U) + 1 to account for an intercept  
term (first entry)
}
\examples{
#### Log-likelihood of an IHG model with no covariate
m<-10; theta<-0.14; n<-300
ordinal<-simihg(n,m,theta)
loglik<-loglikIHG(ordinal,m,param=theta)
##################################
#### Log-likelihood of a IHG model with covariate 
data(relgoods)
attach(relgoods)
m<-10
naord<-which(is.na(HandWork))
nacov<-which(is.na(Gender))
na<-union(naord,nacov)
ordinal<-HandWork[-na]; U<-Gender[-na]
nu<-c(-1.55,-0.11)     # first entry: intercept term
loglik<-loglikIHG(ordinal,m,param=nu,U=U)
}
\seealso{
\code{\link{GEM}}, \code{\link{logLik}}
}
\keyword{htest}

